src/set_todays_words/requirements.txt
google-cloud-firestore==2.7.3
google-cloud-pubsub==2.13.11